adjustments after fixing conflicts with Lucano's work
