+++ clients/admin-ui/src/features/common/nav/v2/routes.ts
@@ -45,7 +45,6 @@ export const TAXONOMY_ROUTE = "/taxonomy";
 export const ABOUT_ROUTE = "/management/about";
 export const CUSTOM_FIELDS_ROUTE = "/management/custom-fields";
 export const EMAIL_TEMPLATES_ROUTE = "/management/email-templates";
-export const MESSAGING_ROUTE = "/messaging";
 export const DOMAIN_RECORDS_ROUTE = "/management/domain-records";
 export const DOMAIN_MANAGEMENT_ROUTE = "/management/domains";
 export const GLOBAL_CONSENT_CONFIG_ROUTE = "/management/consent";
